Allow library alias using AS instead of WITH NAME
Currently, the plugin regards a reference to a library import with `LibraryName    AS    Library` as an error (ie. Library.Keyword will produce a syntax error). However, AS is now the preferred way since RF 6.0, see https://robotframework.org/robotframework/latest/RobotFrameworkUserGuide.html#setting-custom-name-to-library.

As far as I understand the code, my changes fix that. Please test it, I am not sure how to actually pack this into a plugin file I could test with.